chore(examples): inline the scene setup into simple-app main.js
The scene is short enough to read top to bottom in main.js, so the
example.js indirection and its helper wrappers are gone. The calls drop
await now that the standalone session is synchronous, and the scalar bar
position coordinate gets the third component it expects.
